+++ src/utils/deprecate.ts
@@ -1,0 +1,5 @@
+export default (warning) => {
+  if (process.env.NODE_ENV !== 'production') {
+    console.warn(warning)
+  }
+}