+++ src/utils/isListener.ts
@@ -1,0 +1,1 @@
+export default (reducer: string): boolean => reducer.includes('/')