+++ test/listener.test.js
@@ -1,0 +1,27 @@
+describe('listener', () => {
+  test('should trigger state changes on another models reducers', () => {
+    const { init, dispatch } = require('../src')
+    
+    const count1 = {
+      state: 0,
+      reducers: {
+        increment: (state, payload) => state + payload,
+      }
+    }
+
+    const count2 = {
+      state: 0,
+      reducers: {
+        'count1/increment': (state, payload) => state + payload
+      }
+    }
+
+    const store = init({
+      models: { count1, count2 }
+    })
+
+    dispatch.count1.increment(1)
+
+    expect(store.getState()).toEqual({ count1: 1, count2: 1 })
+  })
+})